local-requirements.txt
alembic==1.13.1
apispec==6.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ruff==0.3.4
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.29
threadpoolctl==3.4.0
typing_extensions==4.10.0
tzdata==2024.1
webargs==8.4.0
Werkzeug==3.0.1
